% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{data}
\name{book_of_mormon}
\alias{book_of_mormon}
\title{The text of the Book of Mormon}
\format{A data frame with 6604 rows and 19 columns:
\describe{
  \item{\code{volume_id}}{Unique id number for the volume}
  \item{\code{book_id}}{Unique id number for the book}
  \item{\code{chapter_id}}{Unique id number for the chapter}
  \item{\code{verse_id}}{Unique id number for the verse}
  \item{\code{volume_title}}{Name of the volume}
  \item{\code{book_title}}{Name of the book}
  \item{\code{volume_long_title}}{Full name of the volume}
  \item{\code{book_long_title}}{Full name of the book}
  \item{\code{volume_subtitle}}{Subtitle of the volume}
  \item{\code{book_subtitle}}{Subtitle of the book}
  \item{\code{volume_short_title}}{Volume abbreviation}
  \item{\code{book_short_title}}{Book abbreviation}
  \item{\code{volume_lds_url}}{Volume abbreviation used at scriptures.lds.org}
  \item{\code{book_lds_url}}{Book abbreviation used at scriptures.lds.org}
  \item{\code{chapter_number}}{Chapter}
  \item{\code{verse_number}}{Verse}
  \item{\code{text}}{Scripture text}
  \item{\code{verse_title}}{Combined book, chapter, and verse (e.g. 1 Nephi 1:1)}
  \item{\code{verse_short_title}}{Abbreviated combined book, chapter, and verse (e.g. 1 Ne. 1:1)}
}}
\source{
\url{http://scriptures.nephi.org/}
}
\usage{
book_of_mormon
}
\description{
A dataset containing the text of the Book of Mormon.
}
\details{
Excludes the following:
\itemize{
  \item Book of Mormon: Title Page (written by Moroni)
  \item Introduction
  \item The Testimony of the Three Witnesses
  \item The Testimony of Eight Witnesses
  \item Testimony of the Prophet Joseph Smith
  \item A Brief Explanation About the Book of Mormon
}
}
\keyword{datasets}
