% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{data}
\name{doctrine_and_covenants}
\alias{doctrine_and_covenants}
\title{The text of the Doctrine and Covenants}
\format{A data frame with 3654 rows and 19 columns:
\describe{
  \item{\code{volume_id}}{Unique id number for the volume}
  \item{\code{book_id}}{Unique id number for the book}
  \item{\code{chapter_id}}{Unique id number for the chapter}
  \item{\code{verse_id}}{Unique id number for the verse}
  \item{\code{volume_title}}{Name of the volume}
  \item{\code{book_title}}{Name of the book}
  \item{\code{volume_long_title}}{Full name of the volume}
  \item{\code{book_long_title}}{Full name of the book}
  \item{\code{volume_subtitle}}{Subtitle of the volume}
  \item{\code{book_subtitle}}{Subtitle of the book}
  \item{\code{volume_short_title}}{Volume abbreviation}
  \item{\code{book_short_title}}{Book abbreviation}
  \item{\code{volume_lds_url}}{Volume abbreviation used at scriptures.lds.org}
  \item{\code{book_lds_url}}{Book abbreviation used at scriptures.lds.org}
  \item{\code{chapter_number}}{Chapter}
  \item{\code{verse_number}}{Verse}
  \item{\code{text}}{Scripture text}
  \item{\code{verse_title}}{Combined book, chapter, and verse (e.g. Doctrine and Covenants 1:1)}
  \item{\code{verse_short_title}}{Abbreviated combined book, chapter, and verse (e.g. D&C 1:1)}
}}
\source{
\url{http://scriptures.nephi.org/}
}
\usage{
doctrine_and_covenants
}
\description{
A dataset containing the text of the Doctrine and Covenants.
}
\details{
Excludes the following:
\itemize{
  \item Official Declaration 1
  \item Official Declaration 2
}
}
\keyword{datasets}
