% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\docType{package}
\name{scriptuRs}
\alias{scriptuRs}
\alias{scriptuRs-package}
\title{The Standard Works of The Church of Jesus Christ of Latter-day Saints}
\description{
This package contains the complete text of the Standard Works of The Church
of Jesus Christ of Latter-day Saints (LDS), including the King James Bible
(Old and New Testaments), the Book of Mormon, the Doctrine and Covenants, and
the Pearl of Great Price, formatted to be convenient for text analysis.
}
